+++ src/tiles/mod.rs
@@ -1,0 +1,15 @@
+pub mod file;
+#[cfg(feature = "ici")]
+pub mod ici;
+pub mod tilemap;
+pub mod tileset;
+pub mod units;
+
+pub mod prelude {
+    pub use crate::tiles::file::*;
+    #[cfg(feature = "ici")]
+    pub use crate::tiles::ici::*;
+    pub use crate::tiles::tilemap::*;
+    pub use crate::tiles::tileset::*;
+    pub use crate::tiles::units::*;
+}